pyproject.toml
[project]
name = "drone-vlm-golden-eval"
version = "0.1.0"
description = "Self-contained drone VLM golden evaluation workshop demo"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "openai>=1.0",
    "pandas>=2.2",
    "pillow>=10.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "matplotlib>=3.7",
    "rich>=13.0",
    "google-genai>=1.0",
    "ipykernel>=6.0",
    "gradio>=6.14.0",
    "gradio-client>=2.5.0",
]

[tool.uv]
package = false
